+++ src/openGL/IndexBuffer.h
@@ -3,14 +3,14 @@
 class IndexBuffer {
 private:
     unsigned int m_RendererID;
-    unsigned int m_Count;
+    size_t m_Count;
 
 public:
-    IndexBuffer(const unsigned int* data, unsigned int count);
+    IndexBuffer(const unsigned int* data, size_t count);
     ~IndexBuffer();
 
     void Bind() const;
     void Unbing() const;
 
-    inline unsigned int GetCount() const { return m_Count; }
+    inline size_t GetCount() const { return m_Count; }
 };
