+++ src/openGL/Renderer.cpp
@@ -26,6 +26,6 @@ void Renderer::Draw(const VertexArray& va, const IndexBuffer& ib,
     va.Bind();
     ib.Bind();
 
-    GLCall(
-        glDrawElements(GL_TRIANGLES, ib.GetCount(), GL_UNSIGNED_INT, nullptr));
+    GLCall(glDrawElements(GL_TRIANGLES, static_cast<int>(ib.GetCount()),
+                          GL_UNSIGNED_INT, nullptr));
 }
