+++ src/openGL/VertexBuffer.cpp
@@ -2,7 +2,7 @@
 
 #include "Renderer.h"
 
-VertexBuffer::VertexBuffer(const void* data, unsigned int count) {
+VertexBuffer::VertexBuffer(const void* data, size_t count) {
     GLCall(glGenBuffers(1, &m_RendererID));
     GLCall(glBindBuffer(GL_ARRAY_BUFFER, m_RendererID));
     GLCall(glBufferData(GL_ARRAY_BUFFER, count * sizeof(float), data,
